+++ CHANGELOG.md
@@ -7,6 +7,7 @@
 - Adding `c::stripeJsTokenMarkup()`.
 - Adding `c::escSq()`.
 - Adding `c::escDq()`.
+- Adding Semantic UI custom build directory `src/client-s/css/semantic`.
 
 ## v170309.60830